requirements.txt
colorama==0.3.7
envoy==0.0.3
gitdb==0.6.4
GitPython==2.0.8
natsort==5.0.1
nose==1.3.7
six==1.10.0
smmap==0.9.0
